+++ src/RhinoInside.Revit.External/DB/Schemas/ParameterGroup.cs
@@ -32,7 +32,9 @@ namespace RhinoInside.Revit.External.DB.Schemas
 
     public static bool IsParameterGroup(string id)
     {
-      return id.StartsWith("autodesk.parameter.group") || id.StartsWith("autodesk.revit.group");
+      return id == string.Empty || // 'Other'
+             id.StartsWith("autodesk.parameter.group") ||
+             id.StartsWith("autodesk.revit.group");
     }
 
 #if REVIT_2021